pyproject.toml
[tool.poetry]
name = "impacthon-server"
version = "0.1.0"
description = ""
authors = ["MisileLaboratory", "snakeblanket"]
license = "MisileLab License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.0"
motor = "3.3.2"
uvicorn = "0.27.0"
websockets = "12.0"
python-dotenv = "1.0.1"
pyjwt = "2.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
